Move TaskVariant to a separate file
