+++ dumper/app/src/main/java/com/google/edwmigration/dumper/application/dumper/connector/snowflake/SnowflakeLiteConnector.java
@@ -54,23 +54,6 @@ public final class SnowflakeLiteConnector extends AbstractSnowflakeConnector
     return ArchiveNameUtil.getFileName(NAME);
   }
 
-  private static class TaskVariant {
-
-    public final String zipEntryName;
-    public final String schemaName;
-    public final String whereClause;
-
-    public TaskVariant(String zipEntryName, String schemaName, String whereClause) {
-      this.zipEntryName = zipEntryName;
-      this.schemaName = schemaName;
-      this.whereClause = whereClause;
-    }
-
-    public TaskVariant(String zipEntryName, String schemaName) {
-      this(zipEntryName, schemaName, "");
-    }
-  }
-
   private void addSqlTasksWithInfoSchemaFallback(
       @Nonnull List<? super Task<?>> out,
       @Nonnull Class<? extends Enum<?>> header,
